# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FNAS-LCD",
    "description": "Latency-Constrained Neural Architecture Search Web Service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/datasets/upload": {
      "get": {
        "summary": "Dataset Upload Form",
        "operationId": "dataset_upload_form_datasets_upload_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Dataset Upload Handler",
        "operationId": "dataset_upload_handler_datasets_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_dataset_upload_handler_datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/datasets/{dataset_id}": {
      "get": {
        "summary": "Dataset Detail",
        "operationId": "dataset_detail_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "summary": "Search Form",
        "operationId": "search_form_search_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Search Run",
        "description": "Submit a job: search + export run in the background (demo: training\nis skipped). Redirects home with a 'submitted' flash; track progress\nand download from the job's page.",
        "operationId": "search_run_search_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_search_run_search_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "Jobs Page",
        "operationId": "jobs_page_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Job Detail",
        "operationId": "job_detail_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/datasets/{dataset_id}": {
      "get": {
        "summary": "Get Dataset",
        "operationId": "get_dataset_api_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Dataset",
        "operationId": "delete_dataset_api_datasets__dataset_id__delete",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/datasets/upload": {
      "post": {
        "summary": "Upload Dataset",
        "description": "Upload a zip/tar archive in ImageFolder format.",
        "operationId": "upload_dataset_api_datasets_upload_post",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Dataset display name",
              "title": "Name"
            },
            "description": "Dataset display name"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_datasets_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search": {
      "post": {
        "summary": "Run Search",
        "description": "Run architecture search under a latency/FLOPs budget.",
        "operationId": "run_search_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "post": {
        "summary": "Create Job",
        "description": "Create and start a training job.",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_dataset_upload_handler_datasets_upload_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "name",
          "file"
        ],
        "title": "Body_dataset_upload_handler_datasets_upload_post"
      },
      "Body_search_run_search_post": {
        "properties": {
          "dataset_id": {
            "type": "integer",
            "title": "Dataset Id"
          },
          "device_type": {
            "type": "string",
            "title": "Device Type"
          },
          "budget": {
            "type": "number",
            "title": "Budget"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "device_type",
          "budget"
        ],
        "title": "Body_search_run_search_post"
      },
      "Body_upload_dataset_api_datasets_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_api_datasets_upload_post"
      },
...
      "DatasetResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "num_classes": {
            "type": "integer",
            "title": "Num Classes"
          },
          "is_upload": {
            "type": "boolean",
            "title": "Is Upload"
          },
          "is_ready": {
            "type": "boolean",
            "title": "Is Ready"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path",
          "num_classes",
          "is_upload",
          "is_ready"
        ],
        "title": "DatasetResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreateRequest": {
        "properties": {
          "dataset_id": {
            "type": "integer",
            "title": "Dataset Id"
          },
          "device_type": {
            "type": "string",
            "title": "Device Type",
            "default": "jetson_gpu"
          },
          "budget": {
            "type": "number",
            "title": "Budget"
          },
          "epochs": {
            "type": "integer",
            "title": "Epochs",
            "default": 100
          },
          "warmup": {
            "type": "integer",
            "title": "Warmup",
            "default": 5
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size",
            "default": 64
          },
          "workers": {
            "type": "integer",
            "title": "Workers",
            "default": 4
          },
          "lr": {
            "type": "number",
            "title": "Lr",
            "default": 0.1
          },
          "gpu": {
            "type": "integer",
            "title": "Gpu",
            "default": 0
          },
          "max_iters": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Iters"
          },
          "structure_str": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure Str"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "budget"
        ],
        "title": "JobCreateRequest"
      },
      "JobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "jobs"
        ],
        "title": "JobListResponse"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "dataset_id": {
            "type": "integer",
            "title": "Dataset Id"
          },
          "dataset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Name"
          },
          "device_type": {
            "type": "string",
            "title": "Device Type"
          },
          "budget": {
            "type": "number",
            "title": "Budget"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "aggregate": {
            "type": "string",
            "title": "Aggregate"
          },
          "epochs": {
            "type": "integer",
            "title": "Epochs"
          },
          "warmup": {
            "type": "integer",
            "title": "Warmup"
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size"
          },
          "workers": {
            "type": "integer",
            "title": "Workers"
          },
          "lr": {
            "type": "number",
            "title": "Lr"
          },
          "gpu": {
            "type": "integer",
            "title": "Gpu"
          },
          "max_iters": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Iters"
          },
          "structure_str": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure Str"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "best_val_acc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Val Acc"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "save_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Save Dir"
          },
          "search_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Score"
          },
          "search_constraint": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Constraint"
          },
          "search_flops": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Flops"
          },
          "search_params": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Params"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dataset_id",
          "device_type",
          "budget",
          "metric",
          "aggregate",
          "epochs",
          "warmup",
          "batch_size",
          "workers",
          "lr",
          "gpu",
          "status"
        ],
        "title": "JobResponse"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
      "SearchRequest": {
        "properties": {
          "dataset_id": {
            "type": "integer",
            "title": "Dataset Id"
          },
          "device_type": {
            "type": "string",
            "title": "Device Type",
            "default": "jetson_gpu"
          },
          "budget": {
            "type": "number",
            "title": "Budget"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "budget"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/SearchResult"
          },
          "device_type": {
            "type": "string",
            "title": "Device Type"
          },
          "budget": {
            "type": "number",
            "title": "Budget"
          },
          "num_classes": {
            "type": "integer",
            "title": "Num Classes"
          }
        },
        "type": "object",
        "required": [
          "result",
          "device_type",
          "budget",
          "num_classes"
        ],
        "title": "SearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}